only have a single noop menu implementation
